src/copy_that/application/sam_segmentation.py
"""
SAM (Segment Anything Model) integration for semantic color segmentation.

Provides:
- Semantic segmentation masks
- Object detection and segmentation
- Color region identification
- Educational framework for advanced vision models
"""

import logging

import numpy as np

logger = logging.getLogger(__name__)


class SAMColorSegmentation:
    """SAM-based color segmentation (framework for FastSAM/SAM integration)"""

    def __init__(self, model_type: str = "fastsam", device: str = "cpu"):
        """Initialize SAM segmentation

        Args:
            model_type: "sam" or "fastsam"
            device: "cpu" or "cuda"

        Note: Actual model loading would happen here in production
        """
        self.model_type = model_type
        self.device = device
        self.model = None
        logger.info(f"SAM segmentation initialized with {model_type} on {device}")

    def load_model(self):
        """Load SAM/FastSAM model

        In production, this would download and load the actual model.
        For now, this is a placeholder for the education framework.

        Installation:
            pip install ultralytics  # For FastSAM
            # or
            pip install segment-anything  # For SAM
        """
        try:
            if self.model_type == "fastsam":
                # from ultralytics import FastSAM
                # self.model = FastSAM("FastSAM-s.pt")
                logger.warning("FastSAM not installed. Install with: pip install ultralytics")
            elif self.model_type == "sam":
                # from segment_anything import sam_model_registry, SamPredictor
                # self.model = sam_model_registry["vit_h"](checkpoint="sam_vit_h_4b8939.pth")
                logger.warning("SAM not installed. Install with: pip install segment-anything")
            else:
                raise ValueError(f"Unknown model type: {self.model_type}")

            logger.info(f"Model {self.model_type} loaded successfully")
        except ImportError as e:
            logger.error(f"Failed to load SAM model: {e}")
            logger.info("Install requirements: pip install ultralytics segment-anything")

    def segment_colors(
        self,
        image: np.ndarray,
        conf: float = 0.4,
        iou: float = 0.9
    ) -> list[dict]:
        """Segment image and identify color regions

        Args:
            image: Input image (RGB)
            conf: Confidence threshold
            iou: IOU threshold for NMS

        Returns:
            List of segmented regions with color information
        """
        if self.model is None:
            return self._fallback_segmentation(image)

        try:
            if self.model_type == "fastsam":
                return self._segment_with_fastsam(image, conf, iou)
            elif self.model_type == "sam":
                return self._segment_with_sam(image)
        except Exception as e:
            logger.error(f"Segmentation failed: {e}")
            return self._fallback_segmentation(image)

    def _segment_with_fastsam(
        self,
        image: np.ndarray,
        conf: float,
        iou: float
    ) -> list[dict]:
        """Segment using FastSAM"""
        # Placeholder for actual FastSAM implementation
        # results = self.model(image, conf=conf, iou=iou)
        # Process results...
        logger.warning("FastSAM integration not fully implemented")
        return self._fallback_segmentation(image)

    def _segment_with_sam(self, image: np.ndarray) -> list[dict]:
        """Segment using SAM"""
        # Placeholder for actual SAM implementation
        # predictor = SamPredictor(self.model)
        # predictor.set_image(image)
        # Process with different prompt strategies...
        logger.warning("SAM integration not fully implemented")
        return self._fallback_segmentation(image)

    @staticmethod
    def _fallback_segmentation(image: np.ndarray) -> list[dict]:
        """Fallback: Simple color-based segmentation using K-means regions

        Used when SAM models are not available
        """
        h, w = image.shape[:2]

        # Simple grid-based segmentation
        grid_size = 4
        region_h = h // grid_size
        region_w = w // grid_size

        regions = []
        region_id = 0

        for i in range(grid_size):
            for j in range(grid_size):
                y1 = i * region_h
                y2 = min((i + 1) * region_h, h)
                x1 = j * region_w
                x2 = min((j + 1) * region_w, w)

                region_img = image[y1:y2, x1:x2]
                mean_color = region_img.reshape(-1, 3).mean(axis=0).astype(np.uint8)

                regions.append({
                    'id': region_id,
                    'mask': np.ones((y2 - y1, x2 - x1), dtype=np.uint8) * 255,
                    'bbox': (x1, y1, x2, x2),
                    'color_rgb': tuple(mean_color),
                    'color_hex': f"#{mean_color[0]:02X}{mean_color[1]:02X}{mean_color[2]:02X}",
                    'area': (x2 - x1) * (y2 - y1),
                    'confidence': 0.5,  # Lower confidence for fallback
                })
                region_id += 1

        return regions

    def get_mask_base64(self, mask: np.ndarray) -> str:
        """Encode segmentation mask as base64 string"""
        import base64

        import cv2

        # Convert mask to PNG
        _, buffer = cv2.imencode('.png', mask)
        mask_b64 = base64.b64encode(buffer).decode('utf-8')
        return f"data:image/png;base64,{mask_b64}"


class SegmentationColorAnalyzer:
    """Analyze colors within segmented regions"""

    @staticmethod
    def analyze_segment_colors(
        image: np.ndarray,
        mask: np.ndarray
    ) -> dict:
        """Analyze color properties within a segmented region

        Args:
            image: RGB image
            mask: Binary segmentation mask

        Returns:
            Dictionary with color statistics
        """
        # Apply mask
        masked_image = image.copy()
        masked_image[mask == 0] = [0, 0, 0]

        # Get pixels in region
        region_pixels = image[mask > 0]

        if len(region_pixels) == 0:
            return {}

        # Calculate statistics
        mean_color = region_pixels.mean(axis=0).astype(np.uint8)
        std_color = region_pixels.std(axis=0)
        min_color = region_pixels.min(axis=0)
        max_color = region_pixels.max(axis=0)

        return {
            'mean_rgb': tuple(mean_color),
            'std_rgb': tuple(std_color),
            'min_rgb': tuple(min_color),
            'max_rgb': tuple(max_color),
            'pixels_count': len(region_pixels),
        }
