pyproject.toml
[project]
name = "copy-that"
version = "1.0.0"
description = "Multi-Modal Token Platform - Transform images, video, and audio into universal design tokens"
authors = [
    {name = "Josh Band", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["design-tokens", "ai", "computer-vision", "multi-modal", "saas"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    # Web Framework (FastAPI for SaaS)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.0.0",  # Production WSGI server
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database (PostgreSQL for production SaaS)
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",  # Async PostgreSQL

    # Authentication & Security (SaaS essentials)
    "python-jose[cryptography]>=3.3.0",  # JWT
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",

    # GCP Services
    "google-cloud-storage>=2.14.0",
    "google-cloud-vision>=3.7.0",
    "google-cloud-aiplatform>=1.42.0",
    "google-cloud-secret-manager>=2.18.0",

    # AI/ML
    "anthropic>=0.18.0",  # Claude API
    "openai>=1.0.0",  # OpenAI GPT-4V API
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "coloraide>=3.0.0",  # Color science library

    # Caching & Queue (SaaS performance)
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",  # Add Redis support
    "python-dotenv>=1.0.0",  # Ensure dotenv is available

    # Monitoring & Observability (SaaS critical)
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "sentry-sdk[fastapi]>=1.40.0",

    # Multi-tenancy
    "python-dotenv>=1.0.0",

    # Utils
    "httpx>=0.26.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",  # For FastAPI testing

    # Code Quality
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Type Stubs
    "types-python-dateutil",
    "types-redis",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/joshband/copy-that"
Documentation = "https://copy-that.dev"
Repository = "https://github.com/joshband/copy-that"
"Bug Tracker" = "https://github.com/joshband/copy-that/issues"

[project.scripts]
copy-that = "copy_that.interfaces.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/copy_that"]

# Ruff Configuration (replaces black, isort, flake8, pylint)
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function call in argument defaults
    "E741",  # ambiguous variable name (l is common in color code)
    "SIM105", # contextlib.suppress suggestion
    "SIM108", # ternary operator suggestion
    "SIM113", # enumerate suggestion
    "C401",  # unnecessary generator
    "C403",  # unnecessary list comprehension
    "ARG001", # unused function argument (common in interfaces)
    "ARG002", # unused method argument (common in interfaces)
    "B904",  # raise from (nice to have but not critical)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__
"tests/**/*.py" = ["ARG", "F841", "B007"]  # Allow unused arguments/variables in tests
"alembic/**/*.py" = ["E402", "ARG"]  # Allow late imports and unused args in migrations
"tests/conftest.py" = ["E402"]  # Allow late imports for model registration
"test_*.py" = ["F841", "B007"]  # Allow unused variables in root test files
"src/copy_that/application/tests/**/*.py" = ["F841", "B007"]  # Allow in app tests

# MyPy Configuration
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=src/copy_that",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src/copy_that"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
